+++ app/Http/Controllers/PublisherController.php
@@ -6,6 +6,11 @@ use Illuminate\Http\Request;
 
 class PublisherController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     // Exibe uma lista de categorias
     public function index()
     {
@@ -16,12 +21,15 @@ class PublisherController extends Controller
     // Mostra o formulário para criar uma nova categoria
     public function create()
     {
+        $this->authorize('create', Publisher::class);
         return view('publisher.create');
     }
 
     // Armazena uma nova categoria no banco de dados
     public function store(Request $request)
     {
+        $this->authorize('create', Publisher::class);
+        
         $request->validate([
             'name' => 'required|string|unique:publishers|max:255',
         ]);
@@ -34,18 +42,22 @@ class PublisherController extends Controller
     // Exibe uma categoria específica
     public function show(Publisher $publisher)
     {
+        $this->authorize('view', $publisher);
         return view('publisher.show', compact('publisher'));
     }
 
     // Mostra o formulário para editar uma categoria existente
     public function edit(Publisher $publisher)
     {
+        $this->authorize('update', $publisher);
         return view('publisher.edit', compact('publisher'));
     }
 
     // Atualiza uma categoria no banco de dados
     public function update(Request $request, Publisher $publisher)
     {
+        $this->authorize('update', $publisher);
+        
         $request->validate([
             'name' => 'required|string|unique:publishers,name,' . $publisher->id . '|max:255',
         ]);
@@ -58,6 +70,7 @@ class PublisherController extends Controller
     // Remove uma categoria do banco de dados
     public function destroy(Publisher $publisher)
     {
+        $this->authorize('delete', $publisher);
         $publisher->delete();
 
         return redirect()->route('publisher.index')->with('success', 'publisher excluída com sucesso.');
